+++ src/main/java/szewek/mcflux/fluxcompat/ForgeEnergyCapable.java
@@ -1,4 +1,4 @@
-package szewek.mcflux.wrapper;
+package szewek.mcflux.fluxcompat;
 
 import szewek.fl.energy.IEnergy;